refactor: :recycle: import from `internals` only, not sub-modules of it
